+++ src/components/takes/TakeEntryDetailsNew.jsx
@@ -1,15 +1,7 @@
-import { useCallback } from 'react';
-
-import OneTwoThreeIcon from '@material-symbols/svg-400/rounded/123.svg';
-import DeleteIcon from '@material-symbols/svg-400/rounded/delete.svg';
-import UploadIcon from '@material-symbols/svg-400/rounded/upload.svg';
-
+import SettingsSceneShotsRenumberButton from '@/components/shots/settings/SettingsSceneShotsRenumberButton';
+import SettingsShotDeleteButton from '@/components/shots/settings/SettingsShotDeleteButton';
+import SettingsShotTakesImportButton from '@/components/shots/settings/SettingsShotTakesImportButton';
 import HorizontallySnappableDiv from '@/libs/HorizontallySnappableDiv';
-import { useMultiFileInput } from '@/libs/UseMultiFileInput';
-import { useTakeExporter } from '@/serdes/UseTakeExporter';
-import { isShotEmpty } from '@/stores/document';
-import { useDocumentStore } from '@/stores/document/use';
-import { useSetUserCursor } from '@/stores/user';
 
 import { getListDecorationStyleByViewMode } from './TakeListViewMode';
 
@@ -30,55 +22,24 @@ export default function TakeEntryDetails({
   shotId,
   viewMode,
 }) {
-  const reorderShots = useDocumentStore((ctx) => ctx.reorderShots);
-  const emptyShot = useDocumentStore((ctx) =>
-    isShotEmpty(ctx, documentId, shotId),
-  );
-  const deleteShot = useDocumentStore((ctx) => ctx.deleteShot);
-  const setUserCursor = useSetUserCursor();
-
-  const exportTake = useTakeExporter();
-  /** @type {import('@/libs/UseMultiFileInput').MultiFileInputChangeHandler} */
-  const onFile = useCallback(
-    function _onFile(files) {
-      for (let file of files) {
-        console.log('[ImportFootage] Imported ' + file.name + ' ' + file.type);
-        exportTake(file, documentId, sceneId, shotId);
-      }
-    },
-    [documentId, sceneId, shotId, exportTake],
-  );
-  const [render, click] = useMultiFileInput('video/*', onFile);
-
   const listDecorationStyle = getListDecorationStyleByViewMode(viewMode);
   return (
     <HorizontallySnappableDiv className={className + ' ' + listDecorationStyle}>
       <>
-        <button
-          className="flex flex-row p-2 ml-2 outline rounded text-gray-400 hover:text-black"
-          onClick={click}>
-          <UploadIcon className="w-6 h-6 fill-current ml-auto" />
-          <span className="mr-auto">Import footage</span>
-          {render()}
-        </button>
-        <button
-          className="flex flex-row p-2 ml-2 outline rounded text-gray-400 enabled:hover:text-black disabled:opacity-30"
-          onClick={() => {
-            if (emptyShot) {
-              setUserCursor(documentId, sceneId, '', '');
-              deleteShot(documentId, shotId);
-            }
-          }}
-          disabled={!emptyShot}>
-          <DeleteIcon className="w-6 h-6 fill-current ml-auto" />
-          <span className="mr-auto">Delete shot</span>
-        </button>
-        <button
-          className="flex flex-row p-2 ml-2 outline rounded text-gray-400 enabled:hover:text-black disabled:opacity-30"
-          onClick={() => reorderShots(documentId, sceneId, true)}>
-          <OneTwoThreeIcon className="w-6 h-6 fill-current ml-auto" />
-          <span className="mr-auto">Renumber shots</span>
-        </button>
+        <SettingsShotTakesImportButton
+          documentId={documentId}
+          sceneId={sceneId}
+          shotId={shotId}
+        />
+        <SettingsShotDeleteButton
+          documentId={documentId}
+          sceneId={sceneId}
+          shotId={shotId}
+        />
+        <SettingsSceneShotsRenumberButton
+          documentId={documentId}
+          sceneId={sceneId}
+        />
       </>
     </HorizontallySnappableDiv>
   );